current implementation suffers from poor modularity score
Added the code to calculate modularity score for the detected
communities and what I found is that for the network ninetriangle.net
the modularity score is 0.23 in my distributed code whereas in the
shared code the modularity score is 0.45.

Going to change few things in the code to see if there comes any
different outcome.